+++ temporal_boost/__init__.py
@@ -1,6 +1,5 @@
 from .core import BoostApp  # noqa: F401
 from .schemas import BoostOTLPConfig  # noqa: F401
 from .logger import BoostLoggerConfig  # noqa: F401
-from .http import BoostHTTPRoute  # noqa: F401
 
 __all__ = ["BoostApp", "BoostLoggerConfig"]
