--- temporal_boost/http/__init__.py
@@ -1,1 +1,0 @@
-from .worker import BoostHTTPWorker, BoostHTTPRoute  # noqa: F401
