+++ temporal_boost/internal/__init__.py
@@ -1,0 +1,1 @@
+from .worker import InternalWorker  # noqa: F401
